Show diagram source code below the Krayon-drawn diagrams
